Improved the central thumb
Changed the ligament cover attachment method to a larger wedge at the
distal end an two bolts.
Reshaped the thumb to be larger, and to better meet the 1st and 2nd
fingers for a pinch and tripod.
Improved the thumb ligament to have all the joint protection and
stabilty features of the fingers.
Also some cleaning up of some unused models in the file