Modified ligament cover to avoid M2 bolts
The ligament cover no longer interferes with the back cover M2 bolts.
The bolts themsleves had to be move 1 mm laterally from the centre of
the hand to avoid the PQ12s that they were aslo brushing. As a result,
the inner palm, the outer palm frame, and the inner palm frame were also
modified.