Improvements to the clip on palm
The outer palm frame was fracturing along layer lines as it was
assembled #49. The sharp edges were rounded to reduce stress
concentration. The part was printed Y-up instead of Z-up to alter the
layer line orientation. The outer frame thickness was increased from 1
mm to 2 mm and 3 mm in the lateral and medial, palmar edges. The mould
was also fixed, now taking into account thumb sweep.