Added platforms for the Ligament cover bolt holes
Putting in the heat fit inserts is easier when going into a flat,
perpendicular face. This will also ensure the standard bolt length will
reach the insert #63. Before, as the insert was going into a slope, it
was unclear where to push it too. This could result in it being too deep
for the short bolt.